Medium: Compute local majorant grid and use it for distance sampling
This commit adds a local majorant grid to the `Medium` class and uses
it to sample distances when available. Majorant grid traversal is done
using a hard-coded DDA implementation taken from 
Nimier-David et al. (2022). This is a preliminary implementation: the 
DDA algorithm should actually be attached to the volume data source, 
since the intent is to vary the traversal implementation depending on 
the volume data source.