Distant sensor tests: Use an analytical rectangle shape where necessary
This commit works around an issue that appeared after the transition of
the rectangle shape implementation to a mesh composed of 2 triangles
(see mitsuba-renderer/mitsuba3#1784).

When relying on the native kd-tree, the mesh intersection routine can
leak some rays intersecting edges. This lead to distant sensor tests
with a point target set to (0, 0, 0) failing. The workaround consists in
using the analytical rectangle shape, which doesn't have an edge at
(0, 0, 0).